+++ app/admin/components/vpManagementDirective.js
@@ -283,6 +283,10 @@
                                 self.activeCP.visibleOnChpl = true;
                             self.activeCP.certDate = new Date(self.activeCP.certificationDate);
                         });
+                    commonService.getCap(self.cpSelect)
+                        .then(function (cap) {
+                            self.activeCP.cap = cap.plans;
+                        });
                 }
             };